- Bump version to 7.000.341 (7.00 beta 2). - Minor code tweaks.
git-svn-id: svn://svn.code.sf.net/p/nikkhokkho/code/trunk@78 ddd8a7e7-04a1-4c0d-ae79-48073a28b335

+++ FileOptimizer/Source/clsUtil.cpp
@@ -706,9 +706,9 @@ bool __fastcall clsUtil::LoadForm(TForm *pfrmForm)
 	pfrmForm->Height = GetIni(pfrmForm->Name.c_str(), _T("Height"), pfrmForm->Height);
 	iWindowState = GetIni(pfrmForm->Name.c_str(), _T("WindowState"), (int) pfrmForm->WindowState);
 	//We do not want to reopen FileOptimizer in minimized state
-	if (iWindowState == wsMinimized)
+	if (iWindowState == (int) wsMinimized)
 	{
-		iWindowState = wsNormal;
+		iWindowState = (int) wsNormal;
 	}
 	pfrmForm->WindowState = (TWindowState) iWindowState;
 	pfrmForm->DefaultMonitor = (TDefaultMonitor) GetIni(pfrmForm->Name.c_str(), _T("DefaultMonitor"), pfrmForm->DefaultMonitor);
